+++ aframe/src/location-based/index.js
@@ -1,0 +1,5 @@
+import './arjs-webcam-texture'
+import './gps-camera'
+import './gps-entity-place'
+import './gps-projected-camera'
+import './gps-projected-entity-place'