+++ aframe/src/system-arjs.js
@@ -25,6 +25,11 @@ AFRAME.registerSystem('arjs', {
             type: 'string',
             default: '',
         },
+        // new video texture mode (location based only)
+        videoTexture: {
+            type: 'boolean',
+            default: false
+        },
         // old parameters
         debug: {
             type: 'boolean',
@@ -94,6 +99,14 @@ AFRAME.registerSystem('arjs', {
 
     init: function () {
         var _this = this
+
+        // If videoTexture is set, skip the remainder of the setup entirely and just use the arjs-webcam-texture component
+        if(this.data.videoTexture === true && this.data.sourceType === 'webcam') {
+            var webcamEntity = document.createElement("a-entity");
+            webcamEntity.setAttribute("arjs-webcam-texture", true);
+            this.el.sceneEl.appendChild(webcamEntity);
+            return;
+        }
 
         //////////////////////////////////////////////////////////////////////////////
         //		setup arProfile
@@ -219,7 +232,7 @@ AFRAME.registerSystem('arjs', {
 
     tick: function () {
         // skip it if not yet isInitialised
-        if (this.isReady === false) return
+        if (this.isReady === false || this.data.videoTexture === true) return
 
         // update arSession
         this._arSession.update()