+++ three.js/src/threex/arjs-context.js
@@ -77,8 +77,7 @@ Object.assign(Context.prototype, THREE.EventDispatcher.prototype);
 
 // default to github page
 Context.baseURL = 'https://ar-js-org.github.io/AR.js/three.js/'
-Context.REVISION = '3.0.2';
-
+Context.REVISION = '3.2.0'
 /**
  * Create a default camera for this trackingBackend
  * @param {string} trackingBackend - the tracking to user